Login Bank (WIP)
These source files make up a new project I am working on, it is a secure login bank. This can be used to store logins locally on a computer, or locally through an SQL server on the computer. At this point it is a work in progress, I still have many features to implement but the skeleton code is all done. I am going to focus on getting a working version of data being stored locally on the machine first, then move on to adding SQL server integration. I will be working on this intermittently outside of course work.